final maintainability refac for this pr
