+++ source/electrical_heating/src/opengeh_electrical_heating/domain/transformations/__init__.py
@@ -1,4 +1,6 @@
 from .common import calculate_daily_quantity
+from .electrical_heating_calculation import calculate_electrical_heating_in_local_time
+from .electrical_heating_filter_unchanged import filter_unchanged_electrical_heating
 from .metering_points import get_joined_metering_point_periods_in_local_time
 from .time_series_points import get_daily_consumption_energy_in_local_time, get_electrical_heating_in_local_time
 
@@ -7,4 +9,6 @@ __all__ = [
     "get_joined_metering_point_periods_in_local_time",
     "get_electrical_heating_in_local_time",
     "get_daily_consumption_energy_in_local_time",
+    "calculate_electrical_heating_in_local_time",
+    "filter_unchanged_electrical_heating",
 ]
